libcamera_app: Add libcamera_app.cpp as the implementation of class LibcameraApp
This gives us quicker compilations for all the libcamera apps.
